extract pipeline component object and missing data error into pipeline_utils
