+++ src/SOPRANO/pipeline_utils.py
@@ -1,0 +1,26 @@
+from SOPRANO.objects import Parameters
+
+
+class MissingDataError(Exception):
+    pass
+
+
+class _PipelineComponent:
+    """
+    Components of the pipeline are designed to follow the pattern:
+
+    Component.apply(params)
+
+    where apply should include the call to check_read()
+    to permit execution.
+
+    Pipeline components should override these methods.
+    """
+
+    @staticmethod
+    def apply(params: Parameters):
+        pass
+
+    @staticmethod
+    def check_ready(params: Parameters):
+        pass
